Unit Tests - Developing ManagerTask + TestUtils
Building out unit tests to setup TaskTideManager. Using states + summary counts as the test measures.

Goal is to let the engine be engine, core-lib models + database nuances, manager for interacting with DB. Though it could end up being package of core, as its not as substantial a contribution to the system as say rest api etc