Function unit testing WorkItemExecutor & WorkItemProcessor
Single test to verify it works, in that ItemTasks are processed concurrently for the Executor, and that the WorkItemProcessor can do this in parallel.

This will in time allow for X WorkItems to have Y tasks from its Workload to be processed concurrently. Where X and Y are user provided, instead of the hardcoded values currently used.

Added an abstract StateObserver implementing the TaskTideWorkerObserver to use the TaskTracker. Allowing the "waitUntilDone" logic, to work as a periodic scan, but also base it's logic on the state of the Processors. Once done separates the concerns of Work processing, from actions taken around them which is handled through the WorkerObservers.

- Dropped the templated packages from the initial setup, and am going to continue with having the core-lib separated from the engine-lib.

- Added initial classes for the TaskTracker package under engine.worker, and added an StateObserver class under the engine.observer.stateobserver package. Abstract because actions taken on them vary depending on WorkItem/ItemTask.

- Added WorkItem orientated methods to the TaskGenerator for creation of example ManagerTasks.

- Added WorkItem orientated methods EngineTestUtils, displaying execution time summary.

- WorkItemObserver bus fixes. One causing it to run indefinitely, wrong boolean logic of lockId